requirements.txt
asgiref==3.8.1
Django==5.2
django-phonenumber-field==8.0.0
djangorestframework==3.16.0
phonenumbers==9.0.2
sqlparse==0.5.3
tzdata==2025.2
